+++ src/main/ts/pane/event-test.ts
@@ -1,16 +1,16 @@
 import {describe, it} from 'mocha';
 
+import Tweakpane from '../index';
 import {TestUtil} from '../misc/test-util';
-import {PlainTweakpane} from './plain-tweakpane';
 
-function createPane(): PlainTweakpane {
-	return new PlainTweakpane({
+function createPane(): Tweakpane {
+	return new Tweakpane({
 		document: TestUtil.createWindow().document,
 		title: 'Title',
 	});
 }
 
-describe(PlainTweakpane.name, () => {
+describe(Tweakpane.name, () => {
 	it('should listen fold event', (done) => {
 		const pane = createPane();
 		pane.on('fold', () => {
