+++ src/main/ts/pane/input-point-2d-test.ts
@@ -6,16 +6,16 @@ import {RangeConstraint} from '../constraint/range';
 import {StepConstraint} from '../constraint/step';
 import {ConstraintUtil} from '../constraint/util';
 import {Point2dPadTextController} from '../controller/value/point-2d-pad-text';
+import Tweakpane from '../index';
 import {TestUtil} from '../misc/test-util';
-import {PlainTweakpane} from './plain-tweakpane';
 
-function createPane(): PlainTweakpane {
-	return new PlainTweakpane({
+function createPane(): Tweakpane {
+	return new Tweakpane({
 		document: TestUtil.createWindow().document,
 	});
 }
 
-describe(PlainTweakpane.name, () => {
+describe(Tweakpane.name, () => {
 	[
 		{
 			expectedClass: Point2dPadTextController,
