+++ src/main/ts/pane/monitor-number-test.ts
@@ -5,17 +5,17 @@ import {MonitorBinding} from '../binding/monitor';
 import {GraphLogController} from '../controller/value/graph-log';
 import {MultiLogController} from '../controller/value/multi-log';
 import {SingleLogMonitorController} from '../controller/value/single-log';
+import Tweakpane from '../index';
 import {TestUtil} from '../misc/test-util';
 import {IntervalTicker} from '../misc/ticker/interval';
-import {PlainTweakpane} from './plain-tweakpane';
 
-function createPane(): PlainTweakpane {
-	return new PlainTweakpane({
+function createPane(): Tweakpane {
+	return new Tweakpane({
 		document: TestUtil.createWindow().document,
 	});
 }
 
-describe(PlainTweakpane.name, () => {
+describe(Tweakpane.name, () => {
 	[
 		{
 			expectedClass: SingleLogMonitorController,
